Rearrange and use convert_keys; Add update_video
