+++ lib/fb/error.rb
@@ -1,0 +1,3 @@
+module Fb
+  class Error < StandardError; end
+end
